Add files via upload
This is the Inoki Academy task for week 2. The task objective was to analyse the No-show dataset to find out if increase in medical scholarships leads to a percentage increase in turning up for medical appointments by gender. This will help the hospital decide the gender to focus their medical scholarshipson.